ISSUE #2 * Move "minos.common.model" contents to "minos.common.model.abc" module.
