+++ minos/common/model/abc/__init__.py
@@ -5,9 +5,13 @@ This file is part of minos framework.
 
 Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
 """
-from .abc import (
+from .fields import (
     ModelField,
+)
+from .model import (
     MinosModel,
+)
+from .types import (
     ARRAY,
     BOOLEAN,
     BYTES,
